fix: prepare for crates.io publication
- Add version specifiers to path deps (zenavif-parse 0.4, zencodec-types 0.1, zenravif 0.1)
- Add exclude list and documentation URL to Cargo.toml
- Migrate zencodec_types::{PixelDescriptor,PixelBuffer} imports to zenpixels
- Fix .channel_type field -> .channel_type() method, layout_compatible signature
- Add #![forbid(unsafe_code)] (relaxed for unsafe-asm and _dev features)
- Tighten module visibility: internal YUV/SIMD modules pub(crate) by default, pub behind _dev
- Only compile yuv_convert_fast (raw unsafe intrinsics) behind _dev feature
- Suppress dead_code warnings in alternative YUV implementation modules
- Fix clippy: deprecated black_box, unnecessary casts, unused enumerate, abs→unsigned_abs
- Remove tracked debug_simd binary (3.8MB ELF), add to .gitignore
- Rewrite README with correct feature names, Quick Start, animation/encoding docs, feature table
- Update lib.rs doc comments with correct feature names and descriptions

+++ examples/inspect_metadata.rs
@@ -31,10 +31,7 @@ fn main() {
             } => {
                 println!(
                     "NCLX: cp={} tc={} mc={} full_range={}",
-                    *color_primaries as u16,
-                    *transfer_characteristics as u16,
-                    *matrix_coefficients as u16,
-                    full_range
+                    *color_primaries, *transfer_characteristics, *matrix_coefficients, full_range
                 );
             }
             zenavif_parse::ColorInformation::IccProfile(data) => {
